Major refactoring
The motivating change is to create a reified, immutable VaseContext
record that has *everything* that Vase needs, completely eliminating all
global or ambient values. All other changes flow from this.

Semantics and usage should be nearly unchanged, although some APIs
needed to change, out of necessesity.

1. When setting up Vase, it is now necessary to create an instance of
VaseContext, put it in an atom, wire it with appropriate master routes
and call `vase/init` on it. This replaces the creating `routes-atom` and
calling `init-descriptor-routes`.

2. The Pedestal server *must* install the
`vase.interceptors/bind-vase-context` interceptor to apply to all
requests to Vase. This be by adding it to the the top level
:interceptors key in the Pedestal service map, or by placing it on the
top-level route in the root-routes provided to the VaseContext
instance. This is necessary in order to give the Vase endpoints access
to the relevant VaseContext.