Code generation should be simpler
Reduce the cognitive load of the actions by returning to a more
macro-like construction. This increases duplication but removes a level
of indirection. The helper functions that built up the action code are
now gone.

A couple of questions for @ohpauleez are left in the code. (Side note:
the ambiguity these questions address was always there, just hidden
under a layer of helper functions.)