Query should allow coercion on multiple parameters
Fixes #46
Closes #47

Previously, a query with multiple parameters would only work if coercion was not needed. With coercion, the generated code would call `coerce-arg-val` with too many arguments.

This change also introduces a new helper for writing query & transaction tests. `test-db-helper` constructs a fresh, in-memory Datomic database with the same schema as `test_descriptor.edn`. We can use it to test the actions in isolation without requiring a fully-connected Vase + Pedestal.